data: relocate to ph2run/
The data subsystem implements HDF5 I/O on top of simul.h5;
it is consumed exclusively by the CLI driver and the
algorithm shims, not by phase2 compute kernels.  Move the
implementation out of phase2/ into ph2run/:

  phase2/data.c           -> ph2run/data.c
  include/phase2/data.h   -> include/ph2run/data.h

Every consumer rebinds its include path.  No behaviour
change; subsequent commits drop the residual phase2-side
calls into data so libphase2.so can link without HDF5.
