+++ include/circ/qdrift.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/prob.h"
 #include "xoshiro256ss.h"
