+++ ph2run/data.c
@@ -26,7 +26,7 @@
 
 #include "log.h"
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/paulis.h"
 #include "phase2/world.h"
 #include <complex.h>
