+++ phase2/state_prep_coeff.c
@@ -20,7 +20,7 @@
 #include "det_small.h"
 #include "log.h"
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/qreg.h"
 #include "phase2/state_prep_coeff.h"
