+++ test/t-data_attr.c
@@ -17,7 +17,7 @@
 
 #include <hdf5.h>
 
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/world.h"
 
 #include "test.h"
