+++ test/t-data_coeff_matrix.c
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/world.h"
 
 #include "test.h"
