+++ test/t-data_hamil.c
@@ -13,7 +13,7 @@
 #include <stdio.h>
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/paulis.h"
 #include "phase2/world.h"
