+++ test/t-data_mpi.c
@@ -28,7 +28,7 @@
 #include "mpi.h"
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/paulis.h"
 #include "phase2/world.h"
