+++ test/t-data_multidet.c
@@ -11,7 +11,7 @@
 #include <stdio.h>
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/world.h"
 
 #include "t-data.h"
