+++ test/t-data_open.c
@@ -1,7 +1,7 @@
 #include "c23_compat.h"
 #include <stdint.h>
 
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/world.h"
 
 #include "t-data.h"
