+++ test/t-ref-bendazzoli.c
@@ -9,7 +9,7 @@
 #include <string.h>
 
 #include "phase2/circ.h"
-#include "phase2/data.h"
+#include "ph2run/data.h"
 #include "phase2/qreg.h"
 #include "phase2/state_prep_coeff.h"
 #include "phase2/world.h"
